Time: 0 ms (100.00%) | Memory: 7.9 MB (72.71%) - LeetSync

+++ 278-first-bad-version/first-bad-version.cpp
@@ -4,9 +4,9 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        long long int left = 1;
-        long long int right = n;
-        long long int mid;
+        long int left = 1;
+        long int right = n;
+        long int mid;
         while(left <= right){
             mid  = (left + right) / 2;
             if(isBadVersion(mid) == false){
